fix query

+++ src/NoteMap.Core/Domain/Elements/Model/Position.cs
@@ -2,8 +2,8 @@
 {
     public class Position
     {
-        public decimal X { get; set; }
+        public double X { get; set; }
 
-        public decimal Y { get; set; }
+        public double Y { get; set; }
     }
 }
